Issue: #

+++ runtime/src/main/java/org/corfudb/runtime/exceptions/TransactionAbortedException.java
@@ -1,8 +1,7 @@
 package org.corfudb.runtime.exceptions;
 
-import org.corfudb.protocols.wireprotocol.TxResolutionInfo;
-
 import lombok.Getter;
+import org.corfudb.protocols.wireprotocol.TxResolutionInfo;
 
 /**
  * Created by mwei on 1/11/16.
@@ -16,16 +15,22 @@ public class TransactionAbortedException extends RuntimeException {
     TxResolutionInfo txResolutionInfo;
 
     @Getter
-    Integer conflictKey;
+    Long conflictKey;
 
+    /**
+     * Constructor.
+     * @param txResolutionInfo transaction information
+     * @param conflictKey conflict key
+     * @param abortCause cause
+     */
     public TransactionAbortedException(
             TxResolutionInfo txResolutionInfo,
-            Integer conflictKey, AbortCause abortCause) {
-        super("TX ABORT " +
-                " | Snapshot Time = " + txResolutionInfo.getSnapshotTimestamp() +
-                " | Transaction ID = " + txResolutionInfo.getTXid() +
-                " | Conflict Key = " + conflictKey +
-                " | Cause = " + abortCause );
+            Long conflictKey, AbortCause abortCause) {
+        super("TX ABORT "
+                + " | Snapshot Time = " + txResolutionInfo.getSnapshotTimestamp()
+                + " | Transaction ID = " + txResolutionInfo.getTXid()
+                + " | Conflict Key = " + conflictKey
+                + " | Cause = " + abortCause);
         this.txResolutionInfo = txResolutionInfo;
         this.conflictKey = conflictKey;
         this.abortCause = abortCause;
